+++ src/asymmetry/gui/panels/fit/single_tab.py
@@ -82,6 +82,7 @@ from .tab_base import (
     FitTabBase,
     _apply_domain_mismatch_warning,
     _fit_curve_sample_count,
+    _fit_curve_time_bounds,
     _fit_domain_mismatch_message,
     _fit_result_is_usable,
     _fit_summary,
@@ -900,17 +901,9 @@ class SingleFitTab(FitTabBase):
         self._record_fit_verdict(result)
 
         param_dict = {parameter.name: parameter.value for parameter in result.parameters}
-        n_samples = _fit_curve_sample_count(
-            self._composite_model,
-            param_dict,
-            float(self._current_dataset.time.min()),
-            float(self._current_dataset.time.max()),
-        )
-        t_fit = np.linspace(
-            self._current_dataset.time.min(),
-            self._current_dataset.time.max(),
-            n_samples,
-        )
+        t_min, t_max = _fit_curve_time_bounds(self._current_dataset)
+        n_samples = _fit_curve_sample_count(self._composite_model, param_dict, t_min, t_max)
+        t_fit = np.linspace(t_min, t_max, n_samples)
         y_fit = self._composite_model.function(t_fit, **param_dict)
         component_curves = self._composite_model.evaluate_components(
             t_fit,
@@ -972,18 +965,10 @@ class SingleFitTab(FitTabBase):
             parameters.add(param)
 
         param_dict = {p.name: p.value for p in parameters}
-        n_samples = _fit_curve_sample_count(
-            self._composite_model,
-            param_dict,
-            float(self._current_dataset.time.min()),
-            float(self._current_dataset.time.max()),
-        )
+        t_min, t_max = _fit_curve_time_bounds(self._current_dataset)
+        n_samples = _fit_curve_sample_count(self._composite_model, param_dict, t_min, t_max)
         # Generate fitted curve for plotting
-        t_fit = np.linspace(
-            self._current_dataset.time.min(),
-            self._current_dataset.time.max(),
-            n_samples,
-        )
+        t_fit = np.linspace(t_min, t_max, n_samples)
         y_fit = self._composite_model.function(t_fit, **param_dict)
 
         component_curves = self._composite_model.evaluate_components(
@@ -1231,13 +1216,9 @@ class SingleFitTab(FitTabBase):
             result.parameters = _shift_rrf_parameters(result.parameters, rrf_offsets, sign=+1)
 
         param_dict = {parameter.name: parameter.value for parameter in result.parameters}
-        n_samples = _fit_curve_sample_count(
-            self._composite_model,
-            param_dict,
-            float(dataset.time.min()),
-            float(dataset.time.max()),
-        )
-        t_fit = np.linspace(dataset.time.min(), dataset.time.max(), n_samples)
+        t_min, t_max = _fit_curve_time_bounds(dataset)
+        n_samples = _fit_curve_sample_count(self._composite_model, param_dict, t_min, t_max)
+        t_fit = np.linspace(t_min, t_max, n_samples)
         y_fit = self._composite_model.function(t_fit, **param_dict)
         component_curves = self._composite_model.evaluate_components(
             t_fit, additive_only=True, **param_dict
@@ -1356,13 +1337,9 @@ class SingleFitTab(FitTabBase):
         self._synchronize_fraction_value_rows()
 
         param_dict = {p.name: p.value for p in result.parameters}
-        n_samples = _fit_curve_sample_count(
-            model,
-            param_dict,
-            float(dataset.time.min()),
-            float(dataset.time.max()),
-        )
-        t_fit = np.linspace(dataset.time.min(), dataset.time.max(), n_samples)
+        t_min, t_max = _fit_curve_time_bounds(dataset)
+        n_samples = _fit_curve_sample_count(model, param_dict, t_min, t_max)
+        t_fit = np.linspace(t_min, t_max, n_samples)
         y_fit = model.function(t_fit, **param_dict)
         component_curves = model.evaluate_components(t_fit, additive_only=True, **param_dict)
         self.fit_completed.emit(result, (t_fit, y_fit), component_curves)